feat: complete BioStructure structural I/O
Port the remaining Gemmi-backed PDB, mmCIF, and mmJSON reader state into the complete BioStructure model. Add the canonical mmCIF writer, output-group controls, structured errors, and Rust/Python public facades. Cover the implementation with pinned Gemmi parity fixtures, focused integration tests, CI wiring, examples, and synchronized public and development documentation.

+++ crates/cosmolkit-core/src/bio/protein.rs
@@ -268,7 +268,7 @@ impl<'a> ProteinAtomRef<'a> {
     }
 
     #[must_use]
-    pub fn position(self) -> Option<[f32; 3]> {
+    pub fn position(self) -> Option<[f64; 3]> {
         self.protein.structure.atom_position(self.atom_id)
     }
 }
